third_party/blink/web_tests/external/wpt/priority-hints/link-attr.tentative.html

<!DOCTYPE html>
<title>Priority Hints - Link element</title>
<meta name="author" title="Dominic Farolino" href="mailto:domfarolino@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<link id=link1 href=resources/stylesheet.css fetchpriority=high>
<link id=link2 href=resources/stylesheet.css fetchpriority=low>
<link id=link3 href=resources/stylesheet.css fetchpriority=auto>
<link id=link4 href=resources/stylesheet.css fetchpriority=xyz>
<link id=link5 href=resources/stylesheet.css>

<script>
  test(() => {
    assert_equals(link1.fetchpriority, "high", "high fetchpriority is a valid IDL value on the link element");
    assert_equals(link2.fetchpriority, "low", "high fetchpriority is a valid IDL value on the link element");
    assert_equals(link3.fetchpriority, "auto", "auto fetchpriority is a valid IDL value on the link element");
    assert_equals(link4.fetchpriority, "auto", "invalid fetchpriority reflects as 'auto' IDL attribute on the link element");
    assert_equals(link5.fetchpriority, "auto", "missing fetchpriority reflects as 'auto' IDL attribute on the link element");
  }, "fetchpriority attribute on <link> elements should reflect valid IDL values");

  const link = document.createElement("link");
  assert_equals(link.fetchpriority, "auto");
</script>
